+++ zoom/_assets/standard_apps/admin/index.py
@@ -19,6 +19,7 @@ import views
 import users
 import groups
 
+
 def log_data(db, status, n, limit, q):
     """retreive log data"""
 
@@ -81,9 +82,11 @@ def activity_panel(db):
         log.path,
         log.timestamp,
         log.elapsed
-    from log left join users on log.user_id = users.id
-    where server = %s
-    and log.status = 'C'
+    from log force index (log_timestamp), users
+    where
+        log.user_id = users.id
+        and server = %s
+        and log.status = 'C'
     order by timestamp desc
     limit 15
     """, host)
@@ -116,12 +119,15 @@ def error_panel(db):
             username,
             path,
             timestamp
-        from log left join users on log.user_id = users.id
-        where log.status in ("E") and timestamp>=%s
-        and server = %s
+        from log force index (log_timestamp), users
+        where
+            log.user_id = users.id
+            and log.status = "E"
+            and server = %s
+            and timestamp >= %s
         order by log.id desc
         limit 10
-        """, today(), host)
+        """, host, today())
 
     rows = []
     for rec in data:
@@ -148,12 +154,15 @@ def warning_panel(db):
             username,
             path,
             timestamp
-        from log inner join users on log.user_id = users.id
-        where log.status in ("W") and timestamp>=%s
-        and server = %s
+        from log force index (log_timestamp), users
+        where
+            log.user_id = users.id
+            and log.status = "W"
+            and server = %s
+            and timestamp >= %s
         order by log.id desc
         limit 10
-        """, today(), host)
+        """, host, today())
 
     rows = []
     for rec in data:
@@ -403,7 +412,6 @@ class MyView(zoom.View):
         """
         return page(content, title='Log Entry', css=css)
 
-
     def configuration(self):
         """Return the configuration page"""
         get = zoom.system.site.config.get
@@ -527,7 +535,7 @@ class MyView(zoom.View):
                     list(
                         (k, v.replace(':', ': ') if len(v) > 160 else v) for k, v in os.environ.items())
                 ),
-                css = """
+                css="""
                     .content table { width: 100%; }
                     .content table td { vertical-align: top; width: 70%; }
                     .content table td:first-child { width: 25%; }
@@ -540,7 +548,4 @@ class MyView(zoom.View):
         return page(load_content('about.md', version=zoom.__version__ + ' Community Edition'))
 
 
-def main(route, request):
-    """main program"""
-    view = MyView(request)
-    return view(*request.route[1:], **request.data)
+main = zoom.dispatch(MyView)